requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
croniter==2.0.5
discord==2.3.2
discord-typings==0.9.0
docker==7.1.0
emoji==2.12.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
Naked==0.1.32
psycopg2-binary==2.9.9
py-cord==2.5.0
pybit==5.7.0
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
shellescape==3.8.1
six==1.16.0
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
yarl==1.9.4
